Port git to ansible
Use ansible to setup git and it's relevant config files. The bash role
has a dependency on this git role. which is kinda weird but ok for now.
Bash set's up the prompt, but the git role has a helper that is required
by the bash prompt. Perhaps latter I'll fix this.

10-git-aliases has been removed because they are no longer used. I am
cleaning up things I never use while working on this project.