Remove ruby version from brash prompt
Knowing the ruby version used to be important when working for Luckie
because it required us to work on maybe 4 or 5 different projects at a
time. This stayed with me when I moved to Gradesfirst and I never
removed the logic even though I was working on a single project.

Now I work on more projects again however they are in different
languages and technologies, for instance python, ruby, go, node,
ansible and puppet. I've decided to simplify my prompt by only having
git status at a glance for now.