Port ack to ansible
Remove ack from the list of homebrew programs since it gets its own role
in this commit. The role will handle the installation of the program.

Remove .bashrc.d/10-ack-environment this was only used to set the
ACK_PAGER environment variable. I learned I can set this in the .ackrc
file and opted to move it there to centralize all the settings in one
place.