Port chruby to ansible
Rename 30-load-ruby-path to 30-load-ruby-manager. Also setting the
default ruby was moved to that file because this repository is going to
become my personal settings. I don't care to manage flexibility for
others.

You have to manually install ruby after this is run. `ruby-install ruby`
is the command requried. perhaps later I'll make it easier to work with.