Move go environment settings to bash environment settings
The environment settings for go do not interfere with anything with go
is not installed, this makes it safe to put in bash environment
settings.

The check for OSX has been removed in 10-environment, because we can use
ansible to install a linux version instead of using logic inside bash.