Remove .bashrc.d and replace with .autoload
.bashrc.d is used to break up the .bashrc file so things could be
configured dynamically. This ended up being a pain in the ass. Now all
of the .bashrc.d files are in the .bashrc file and a new .autoload
directory exists to put bash libraries in. .autoload should only contain
functions necessary to the .bashrc file. Nothing in the .autoload
directory should be used to configure the shell, only expose functions
for .bashrc to use.