Force tmux command line to respect readline
When typing a command using <Tmux Key>-: it pulls up a tmux command
line. It normally defaults to emacs which behaves similar to readline
and allows you to use readline shortcuts. However, if the environment
variables EDITOR or VISUAL are setup to use vim it then defaults to vi.

Change it to behave exactly how we expect it to by always setting
it to emacs no matter what.